requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
alabaster==0.7.12
    # via sphinx
anyio==3.4.0
    # via starlette
appdirs==1.4.4
    # via scanapi
asgiref==3.4.1
    # via uvicorn
babel==2.9.1
    # via sphinx
bcrypt==3.2.0
    # via passlib
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
charset-normalizer==2.0.9
    # via requests
click==7.1.2
    # via
    #   -r requirements.in
    #   scanapi
    #   uvicorn
colorama==0.4.4
    # via
    #   sphinx
    #   uvicorn
cryptography==36.0.0
    # via python-jose
curlify2==1.0.1
    # via scanapi
docutils==0.17.1
    # via
    #   sphinx
    #   sphinx-rtd-theme
ecdsa==0.17.0
    # via python-jose
fastapi==0.69.0
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
h11==0.12.0
    # via uvicorn
httptools==0.2.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   requests
imagesize==1.3.0
    # via sphinx
jinja2==2.11.3
    # via
    #   scanapi
    #   sphinx
markupsafe==2.0.1
    # via jinja2
mongoengine==0.23.1
    # via -r requirements.in
packaging==21.3
    # via sphinx
passlib[bcrypt]==1.7.4
    # via -r requirements.in
psycopg2==2.9.2
    # via -r requirements.in
psycopg2-binary==2.9.2
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==1.8.2
    # via fastapi
pygments==2.10.0
    # via sphinx
pymongo==3.12.2
    # via mongoengine
pyparsing==3.0.6
    # via packaging
python-decouple==3.5
    # via -r requirements.in
python-dotenv==0.19.2
    # via uvicorn
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
pytz==2021.3
    # via babel
pyyaml==5.4
    # via
    #   -r requirements.in
    #   scanapi
    #   uvicorn
requests==2.26.0
    # via
    #   curlify2
    #   responses
    #   scanapi
    #   sphinx
responses==0.12.1
    # via curlify2
rsa==4.8
    # via python-jose
scanapi==2.6.0
    # via -r requirements.in
six==1.16.0
    # via
    #   bcrypt
    #   ecdsa
    #   python-multipart
    #   responses
sniffio==1.2.0
    # via anyio
snowballstemmer==2.2.0
    # via sphinx
sphinx==4.3.1
    # via
    #   -r requirements.in
    #   sphinx-rtd-theme
sphinx-rtd-theme==1.0.0
    # via -r requirements.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlalchemy==1.4.27
    # via -r requirements.in
starlette==0.15.0
    # via fastapi
typing-extensions==4.0.1
    # via pydantic
urllib3==1.26.7
    # via
    #   requests
    #   responses
uvicorn[standard]==0.15.0
    # via -r requirements.in
watchgod==0.7
    # via uvicorn
websockets==10.1
    # via uvicorn
wheel==0.37.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
